requirements.txt
openclaw-sdk>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
akshare>=1.14.0
tushare>=1.4.0
backtrader>=1.9.78
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
scikit-learn>=1.4.0
scipy>=1.12.0
statsmodels>=0.14.0
ta-lib>=0.4.28
pydantic>=2.0
fastapi>=0.109.0
uvicorn>=0.27.0
loguru>=0.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiohttp>=3.9.0
redis>=5.0.0
apscheduler>=3.10.0
